Fix a getting bundle. Redundant casting Double to Int in Recipe pojo class has been removed. 2 method added for getting lists from Recipe and setting Lists of Steps and Ingredients in the Steps Activity

+++ app/src/main/java/pl/michaldobrowolski/bakingapp/api/model/pojo/Ingredient.java
@@ -49,7 +49,7 @@ public class Ingredient implements Parcelable {
     }
 
     private Ingredient(Parcel in) {
-        this.mQuantity = ((int) in.readValue((int.class.getClassLoader())));
+        this.mQuantity = (Double) in.readValue(int.class.getClassLoader());
         this.mMeasure = ((String) in.readValue((String.class.getClassLoader())));
         this.mIngredient = ((String) in.readValue((String.class.getClassLoader())));
     }
